requirements.txt
thordata-sdk>=1.8.4
python-dotenv>=1.0.0
